+++ docs/changes.rst
@@ -4,6 +4,9 @@ Changelog
 11.0.2 (unreleased)
 -------------------
 
+- Fix hidden ``oira_type`` on edit views.
+  [maurits]
+
 - navtree service: filter out training questions.
   [maurits]